Upgrade to Kotlin 2.3.0

+++ orx-jvm/orx-panel/src/main/kotlin/org/openrndr/panel/elements/Slider.kt
@@ -308,7 +308,7 @@ fun Slider.bind(property: KMutableProperty0<Double>) {
                     if (property.get() != currentValue) {
                         val lcur = property.get()
                         currentValue = lcur
-                        value = lcur.toDouble()
+                        value = lcur
                     }
                 }
                 update()
